4차 Commit
delete/patch 권한 설정
전체 질문리스트 조회(페이징)

+++ server/overflow/src/main/java/com/preproject/overflow/question/service/QuestionService.java
@@ -55,7 +55,6 @@ public class QuestionService {
     public Question updateQuestion(Question question, Member loggedInMember) {
         Question findQuestion = questionRepository.findByQuestionId(question.getQuestionId());
 
-        // Check if the logged-in user's memberId matches the author's memberId
         if (findQuestion.getMember().getMemberId().equals(loggedInMember.getMemberId())) {
             Optional.ofNullable(question.getText())
                     .ifPresent(text -> findQuestion.setText(text));
@@ -65,6 +64,7 @@ public class QuestionService {
             throw new BusinessLogicException(ExceptionCode.MEMBER_PERMISSION_DENIED);
         }
     }
+
     public Question voteQuestion(Question question, Boolean vote) {
         Question findQuestion = questionRepository.findByQuestionId(question.getQuestionId());
         if (vote.equals(true)) {
@@ -129,12 +129,15 @@ public class QuestionService {
         return questionPage;
     }
 
-    public void deleteQuestion(Long questionId) {
+    public void deleteQuestion(Long questionId, Member loggedInMember) {
         Question question = findVerifyQuestion(questionId);
+
+        if (!question.getMember().getMemberId().equals(loggedInMember.getMemberId())) {
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PERMISSION_DENIED);
+        }
         deleteTag(question);
         //downQuestionCount(question);
         questionRepository.delete(question);
-
     }
 
     public void createTag(Question question) {
